- Adding a timer that clears the aria-live region a short time after rendering text in that region. This prevents the div from being accessible via VO after the announcement finishes. - Moving accessibility announcements to a separate component.
